requirements.txt
lxml==5.3.0
opcua==0.98.13
paho-mqtt==2.1.0
protobuf==5.29.3
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
